Move Dispatcher Util class used by view tests
