+++ test/DynamoCoreWpfTests/AnnotationViewTests.cs
@@ -11,6 +11,7 @@ using Dynamo.Utilities;
 using Dynamo.Views;
 using DynamoCoreWpfTests.Utility;
 using NUnit.Framework;
+using SystemTestServices;
 
 namespace DynamoCoreWpfTests
 {
