+++ test/DynamoCoreWpfTests/DynamoTestUIBase.cs
@@ -20,6 +20,7 @@ using DynamoCoreWpfTests.Utility;
 using DynamoShapeManager;
 using DynamoUtilities;
 using NUnit.Framework;
+using SystemTestServices;
 using TestServices;
 
 namespace DynamoCoreWpfTests
