+++ test/DynamoCoreWpfTests/HomePageTests.cs
@@ -8,10 +8,10 @@ using Dynamo.Graph.Workspaces;
 using Dynamo.UI.Controls;
 using Dynamo.UI.Views;
 using Dynamo.ViewModels;
-using DynamoCoreWpfTests.Utility;
 using Microsoft.Web.WebView2.Core;
 using Microsoft.Web.WebView2.Wpf;
 using NUnit.Framework;
+using SystemTestServices;
 
 namespace DynamoCoreWpfTests
 {
