+++ test/DynamoCoreWpfTests/PreferencesGeometryScalingTests.cs
@@ -1,15 +1,13 @@
 using System.Collections.Generic;
 using System.Windows;
 using System.Windows.Controls.Primitives;
+using Dynamo.Graph;
 using Dynamo.Graph.Nodes;
-using Dynamo.Wpf.Views;
-using DynamoCoreWpfTests.Utility;
-using NUnit.Framework;
 using Dynamo.ViewModels;
 using Dynamo.Views;
-using Dynamo.Graph;
-using Dynamo.Graph.Workspaces;
-using Dynamo.Models;
+using Dynamo.Wpf.Views;
+using NUnit.Framework;
+using SystemTestServices;
 
 namespace DynamoCoreWpfTests
 {
