+++ test/DynamoCoreWpfTests/SplashScreenTests.cs
@@ -1,15 +1,11 @@
-
-
-using Dynamo.Configuration;
-using Dynamo.Models;
-using DynamoCoreWpfTests.Utility;
-using DynamoUtilities;
-using NUnit.Framework;
 using System;
 using System.Diagnostics;
 using System.Runtime.InteropServices;
 using System.Threading;
 using System.Threading.Tasks;
+using DynamoUtilities;
+using NUnit.Framework;
+using SystemTestServices;
 
 namespace DynamoCoreWpfTests
 {
