+++ test/DynamoCoreWpfTests/UnitsUITests.cs
@@ -1,17 +1,17 @@
+using System;
+using System.Collections.Generic;
+using System.Linq;
 using System.Xml;
+using CoreNodeModels;
 using Dynamo.Graph;
+using Dynamo.Graph.Nodes;
 using Dynamo.Models;
 using Dynamo.Tests;
 using Dynamo.Utilities;
-using DynamoCoreWpfTests.Utility;
 using NUnit.Framework;
+using SystemTestServices;
 using TestUINodes;
 using UnitsUI;
-using System.Linq;
-using CoreNodeModels;
-using System.Collections.Generic;
-using System;
-using Dynamo.Graph.Nodes;
 
 namespace DynamoCoreWpfTests
 {
