+++ test/Libraries/SystemTestServices/DispatcherUtil.cs
@@ -1,13 +1,8 @@
 using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Security.Permissions;
-using System.Windows.Threading;
 using System.Threading;
-using NUnit.Framework;
+using System.Windows.Threading;
 
-namespace DynamoCoreWpfTests.Utility
+namespace SystemTestServices
 {
     public static class DispatcherUtil
     {
