+++ test/VisualizationTests/HelixWatch3dViewModelPrimitiveTests.cs
@@ -4,15 +4,12 @@ using System.Windows.Media;
 using System.Windows.Media.Media3D;
 using Dynamo.Graph;
 using Dynamo.Graph.Workspaces;
-using Dynamo.Selection;
 using Dynamo.Wpf.ViewModels.Watch3D;
-using DynamoCoreWpfTests.Utility;
 using HelixToolkit.SharpDX.Core;
 using HelixToolkit.Wpf.SharpDX;
 using NUnit.Framework;
 using SharpDX;
-using Watch3DNodeModelsWpf;
-
+using SystemTestServices;
 namespace WpfVisualizationTests
 {
     class HelixWatch3dViewModelPrimitiveTests : WpfVisualizationTests.VisualizationTest
